Adding neutron tests to skip file
There were an increase in the amount of tests running, with the
addition of project specific tests via tempest plugins.
These tests weren't properly tested and require further investigation
to check whether is a bug in the package, tempest is properly configured
to run these tests, or even if the test should be running.
For now, we are just skipping these tests.

Change-Id: I7ee1be72d7c93f38ee5158cb8c7b3da0d362939c