Spliting tempest validate and install role
Splitting the validate-yaml in two: install_skiplist and
validate_yaml. Also renaming the role to use underscore '_'
instead of dash '-' becase it will be required in future
ansible version.

The reason to this split is to reuse these roles when start
to use it in other installers, instead of reinvent the
wheel.

Change-Id: I2a736723087016db3eebf20cea48fadb658ba436