Add addtest command to tempest-skiplist
This patch adds the addtest command to tempest-skiplist in order to
avoid mistakes like duplicated entries, invalid yaml file format, add
namespace instead of full test path. The idea here is to help ruck and
rovers to add tests to skiplist, without the need to edit long yaml
file, and be able to quickly select tests under a specific namespace to
be added without all the boilplate of create entries for each test.

Change-Id: Ibedd2c1307940f01229ecce2a683fdfd51dc612e