Update logic on allowed list
Allow both group and job to be parsed. The job will have precedence over
group, which means, if a job has a group configured, but in the
allowed_list.yml file have also that specific job with a specific set of
tests, the matching job will be choose instead of group. This will allow
us to configure the group for all jobs, without the need of too much
verbosing on the yaml file, as well as specify tests to be executed per
job.
Also, job is the only requried option, group can be optional, we will
always know the job name in zuul, but group might not be used.

An example:
All featureset 052 based job will run tempest minimum scenario tests,
however, we can set to scenario002 job run a completely different set of
tests.

Change-Id: I0677125c08ae1a7a43f22e10be6e0319afbec56a