Fix list of releases for a some tests
The releases key is a list of releases. Having more than one releases
key for one test, means the other releases keys will be overwriten with
the latest one. Unfortunatelly, the validation doesn't caught that and
so far, seems to be impossible to find out since the yaml.safe_load
overwrite it by default.

Change-Id: I7aac4efae5b77f0b0b3a713de8e945e83d1cf29b