Add new _nodeps targets for test runs
On some environment, we may not be able to install dependencies as part
of the test run. Providing molecule_nodeps and pre_commit_nodeps target
allows to ensure we can prepare the env in a step, and run the tests in
a second step, without the need to get a working `sudo` or related
commands.