Provides our own CI container
In order to ensure we have all the needed pieces in the container for
Prow, we provide our own container, with everything installed directly
from the Containerfile.ci

In openshift-ci, the container may be built using one of the parameters
pointing to the Containerfile directly, and we should be good.

The container did successfuly build locally, and running some "make"
command was also working fine.

Note that, by default, the container passes USE_VENV=no since it's all
ready with system-wide installed libraries and dependencies.