Add some new targets to Makefile, add README.md
The README is still incomplete, but provides some hints as to how to
contribute.

The Makefile targets are mostly for the devs, since they allow to run
the tests in a container.

The run-local-test script is also renamed with a more relevant name.

Note that, at this point, molecule tests aren't working. We need a first
role to validate the whole chain. The run_molecule script ensures we
have some content to test, and exits early if needed.