Modify how molecule is running
In order to run molecule on Prow, we need to get rid of the "podman"
driver, and use "delegate" instead. In addition, we're using
"localhost", since it's already running in a container, so we have to
ensure the ansible_connection is set to "local" instead of "ssh".

This leads to a third configuration file for molecule.

The _skeleton_role_ molecule.yml is also updated accordingly, so that
new roles will benefit of the defaults set in the config files.