Run molecule only against modified role(s)
Using some trick, we may be able to know what role(s) were modified in
the PR. This patch enables this automated filtering.

It also open a wide door for external reuse of the Makefile and scripts:
if you have your own repository, you should be able to clone this one
in some nearby path, and then run the following:
```Bash
make run_ctx_molecule \
        MOLECULE_CONFIG=.config/molecule/config_local.yml \
        ROLE_DIR=../my_repo/roles
```

Note: relative path must be used (for now).

This patch also takes care of an issue detected in Prow env, where the
collections are installed in /root/.ansible, and are thereofe not
accessible from the molecule run.
So if we don't use the venv (i.e. we have root), let's install the
collections in a known path, so that they should be reachable by the job
even if it runs without root.