Enforces some good practices in the repository
The PULL_REQUEST_TEMPLATE.md will be displayed upon PR, in order to
ensure contributors are meeting some of the important requirements.

The CONTRIBUTING.md will be shown upon contributor's first PR - and is
reachable from the project root in case anyone needs some "how to start"
documentation.

This patch also injects some new files to the _skeleton_role_ to make it
more compliant with the expressed rules.