Stop relying on galaxy.ansible.com service
With the amount of failures of the service, we can't realy rely on it
anymore. Using plain git as sources for
`ansible-galaxy collection install` should make molecule runs more
stable.

If needed, we may also pin a collection to a specific version, just by
adding "version: foo" to the entry.