Move copy role for use w/ podified containers
Add the copy_container role from ci-config so that itcan be modified
and used with the podified container build/test/push workflow.

In addition, this role has no real molecule tests, since it's mostly
running go scripts. In order to avoid crashing the molecule tests, it
therefore implements a kind of noop check.