Ensure our container is able to mimic Prow env
Container running in Prow CI aren't providing root access. We therefore
need to 100% mimic their usage.

In order to do so, the Containerfile.test is now creating a "prow" user
with its home located in / (like in prow), and without any privileges.

Note that molecule isn't running on Prow for the very reason of missing
privileges escalation. For that specific use, we're therefore passing
"--user root" in order to keep the full accesses we need.

This patch also corrects some issues introduced in the past that weren't
properly raised, and sets some more verbosity for the ansible-test runs.