Modify how molecule is launched
Running from within a container is nice, but we then lack the molecule
logs. So instead of running in the container, let's leverage the
virtualenv and related so that we can properly run on the host.

We also collect logs - that step was missing and we had only the default
contentt

Resolves #44