Correct git source for repo-setup, improve directories
The repo-setup project has been moved to OKO organisation - this patch
ensure we won't lose it.

After some more thoughts, the git clone for repo-setup souldn't be at
the root of the output directory, but in the "tmp" one. This patch
reflects this change.

Also, this patch ensures we get the directories, even if we don't run
the ci_setup role beforehand.