Some more changes in the way container is built
It seems buildah is doing things in a different way with regards to the
"COPY" path related. podman wasn't pushing the whole content, it was
stuck in the "containerfiles" directory.

The Containerfile.test also has some changes in order to correct some of
the order - mostly ensuring we're in the right location before doing
anything.

Those changes are needed for at least pre-commit check - for some
reasons, it failed with the last modifications done to that whole
process.