Implement hooks in ci-framework
Hooks are needed in order to allow users to side-load content in the
deployment process.

The `deploy-edpm.yml` playbook has now a first structure, allowing to
understand better how hooks may be leveraged.

Of course, some pre/post hooks are kind of "the same", since they are
following each others without any other tasks in-between, but getting a
clean name allows to know a bit more about the actual purpose of the
hook in question.

The USAGE.md file has been updated in order to show a bit more the
potentials for hooks.