Adds initial implementation of operator_build role
This patch adds initial implementation of a role to build
OpenStack operators.
The role can build multiple operators, when provided their 'name'
and 'src' code directory.
When testing from a pull request, in order to properly build the
meta-operator, the user needs to provide the PR owner and SHA, to
be used on meta-operator go.mod.