Add libvirt_manager role
This patch adds the mentioned role, and also changes the molecule job in
order to use the "base-crc" job as a parent. This will ensure proper
access to an OpenShift env when needed.

It also correct a typo in the test_deps role, apparently a leading "/"
was missing and wasn't discovered until now.

Co-Authored-By: Cedric Jeanneret <cjeanner@redhat.com>