Better splitting of the mighty edpm-deploy.yml playbook
Splitting it like that would allow to get a cleaner overview of the file
content, allowing an easier maintenance later.

This patch also introduces some more top-level parameters, allowing to
activate or not RHOL/CRC and Libvirt supports.

We also need to properly point to the roles location in order to consume
playbooks from a location that doesn't match Ansible auto-discovery