Modify ci_make wrapper
1. Generate the reproducer under any conditions
Until now, the reproducer script was generated if and only if the
"command" entry was available in the community.general.make output.
This patch allows to generate the reproducer as a plain JSON file
listing the parameters that were passed to the community.general.make
module.
Note that this should be temporary, until the collection has a new
release embedding the needed change.

2. Generate logs
Until now, no log were available, unless there was a failure. In this
condition, the log was "lost" in ansible logging.
This patch corrects this issue, by ensuring a proper logging file is
created with the same name as the reproducer script, in the "logs"
directory we create from the ci_setup role.

3. Ensure we get the "chdir" in the reproducer content
Until now, the reproducer scripts were missing the "chdir" content. The
community.general.make module relies on some ansible internal
"run_command". This method takes the "chdir" as a plain parameter,
therefore the module isn't building and `-C <chdir>` command.
This patch ensures we mimic the module, by adding the needed `pushd` and
`popd`.

4. Extends ansible-test run
Until now, we didn't really take care of the generated content.
This path ensures the generated content matches the expected checksum -
using checksum is easier than matching plain text in Ansible.