Add environment exports to ci_make reproducer
Until now, all of the environment was discarded from the reproducer
scripts. This would lead to errors when people want to re-run the make
command using the script, since they'd miss most of the important
content.

This patch ensures we properly export the environment passed to the
ci_make task.