Refactore once more the main edpm-deploy.yml playbook
This refactoring should allow a cleaner layout, with a better
understanding of the whole thing.

Calling sub-playbooks with indexes directly shows the proper inclusion
order in case people want to get their own playbook based on the onces
we provide.

Doing so, we also ensure hooks are properly called in the right "step",
even if we build our own custom playbook.