Consolidate run_hook "playbook" tasks
Until now, there was no way to avoid a playbook hook to run all the
time.

This patch also allows to pass custom Ansible configuration file. By
default, it will use the one located at the root of the project,
allowing hooks to know about our custom roles and plugins.

There was also a risk we'd miss the proper outputs in case of a
crash - so now we should get the logs even if the hook fails, presented
in the proper way.