Added playbook to dump zuul vars
In end-to-end job, we run nestd ansible-playbook in zuul
jobs.

Developers defined the vars in Zuul job definition and it
needs to be passed to nested plays so that we can run the
tasks based on Zuul job vars.

dump_zuul_vars.yml playbook dumps all the vars in the
zuul_inventory.yml file. It will be consumed in the
nested playbook.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>