Operator build plugin now outputs a results dictionary
To be able to fecth the operators build in the operators build step
the role should pass to the next steps, mainly the deploy one, the
images that were build.

As this role can add useful information for further usage like the
git hash of each build operator the dictionary can be used for other
fields in the future.