Added content provider base job
It will be used to build operator images and push it to the
local registry.

The content provider job will be paused to serve as registry
and the dependent job will recieve the content provider
registry ip address and from that we can pull the image.

This patch also other changes:
- Adds var to support local registry in build_operators role
  - similar var is added in meta operator role[1]
- since the content provider job will run on other projects. It
  adds static path for ci-framework repo.
- It also fixes the insecure registry setting for registry_deploy
  role.

[1]. https://github.com/openstack-k8s-operators/openstack-operator/pull/287

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>