New role: edpm_prepare
This role wraps all of the preparation steps/make targets in order to
get the dataplane ready for the actual EDPM deployment.

The goal here is to make a modular role, calling the install_yamls_makes
tasks in a way we may toggle various switches, or even inject operators
early in the game.