Inject CR snippet for generated computes
Current dataplane-operator has a "hard" limit of 2 nodes. The Framework
doesn't have such a limitation, so we need to make sure the extra nodes
are known during the deploy.

This patch will push extra Custom Resources files in the right location
in order to push the extra computes in place.

We also have to update a bit how molecule tests are being prepared, by
ensuring we get a tiny subset of the cifmw_install_yamls_defaults
content.

Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/229