New hook: deploy ceph using install_yamls target
Following these notes[1] we're able to get a toy ceph. The generated
secret will then be available as `cifmw_ceph_uuid` within the main play.

There are many comments in the hook itself, as well as the ci.yml
"scenario" in order to explain how things are working together.

This patch also modify a bit how the end-to-end job calls the
deploy-edpm.yml playbook, in order to leverage the provided inventory,
so that it mimics in a far better way the usage.

As a side note, ci_make is still showing some issues when it comes to
parsing the "environment" parameter we pass to the task...

[1] https://github.com/fultonj/zed/tree/main/edpm