Change a bit the OWNERS
Less approvers, more reviewers - it should make the overall review
process better and stronger.

Approvers aren't in the reviewer listing in order to ensure we get
2-eyes control on the patches.

In order to be added to the "approvers" list, a reviewer must have a
good "body count" - contributions and reviews.

Some changes will also occurs on the openshift/release project
configuration in order to make things clearer.