Return cifmw_operator_build_output var to child job
The child job depends on content provider. The content provider
needs to return cifmw_operator_build_output var to the child job
so that child job should consume the proper operator images
to do the deployment.

This patch also removes the hardcoded content provider ip
and replace it with var. It also moves the tasks to discover
the content provider ip in pre and pass it as a var to
the inner ansible call.

Note: It also fixes the operator_build role to include
the operator image during make_bundle.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>