Developper entry point for easy usage on local HW
This patch introduces a new playbook and Makefile target allowing for
developpers to easily bootstrap a VM.
That VM, once running, will be ready to consume the ci-framework,
leveraging nested virtualization (must be enabled) in order to run the
CRC VM as well as the wanted amount of compute nodes.

Note that the Makefile target allows to pass any ansible-playbook option
you want. One of the "must have" is probably the "--ask-become-pass",
allowing to provide your password for the couple of "become" actions
(mostly package installation).

We also introduce a new zuul job in order to ensure the bootstrap is
correctly done with the new role. It's running if and only if we edit
the specific role content.