Dynamically generate sanity-check ignore file based on Ansible version
ansible-test uses a fixed pattern to fetch the ignore file to use for
sanity checks, based on it's own version. This commit copies/appends
the content of the ansible ignore.txt file to the collections ignore
one to a ignore file with the proper filename based on the installed
ansible version.