Update mol test playbook for consuming it in edpm-ansible jobs
In order to use ci-framework molecule base zuul jobs in edpm-ansible
project, We need to make the roles path and molecule config path
configurable using ansible vars in molecule-test.yml playbook.

It will help to override the roles_path with edpm_ansible roles
and test it via zuul jobs.

Here is the edpm-ansible pr [1] which uses the same.

[1]. https://github.com/openstack-k8s-operators/edpm-ansible/pull/154
Signed-off-by: Chandan Kumar <raukadah@gmail.com>