Pass additional make_edpm_deploy_env var to inner ansible
https://github.com/openstack-k8s-operators/edpm-ansible/pull/161 adds
the EDPM based zuul job. We build the ansibleee-runner image in the
EDPM zuul job itself and RDO quay registry. This image needs to be
passed to the deployment.

In order to do that We need to set DATAPLANE_RUNNER_IMG install_yamls makefile
var to manipulate the ansibleee-runner image.

This patch adds a conditional to check for make_edpm_deploy_env var set
in pre-playbook and pass it to inner ansible-playbook call.

It will help to test the deployment.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>