Integrate role README in sphinx doc
This patch ensures we'll get the README.md included in sphinx. It
therefore ensures the existing content is consistent across the various
roles, and the "default" README is updated accordingly so that new roles
will get the right format.

Since we're able to build the doc locally for testing/preview, this
patch also adds a .gitignore in order to avoid pushing either the built
doc, or the virtual environment.

We also ensure doc is able to build with a dedicated job, triggered only
when README or docs/ are edited.

Since cifmw-doc isn't linked to ansible, we have to modify a bit the way
collect-logs.yml post-run is gathering data.