Do not kick any molecule nor e2e job on .md editions
Until now, even if we were only editing .md files (aka doc), we would
end with at least the end-to-end-nobuild job running and, if we were
editing a role README.md, we'd see its molecule job being kicked.

This patch ensures zuul won't be triggered on .md nor on /docs/
modifications, meaning less load and time on zuul infra.

Still, in order to get proper feedback from the bot, the "noop" job is
now added, and should be kicked even if nothing is showing up. It's
defined in rdoproject directly, so no need to re-define it here.