Initial relase of the openshift_setup role
This is the first release of the openshift_setup role. This role is
meant to be used in the infra phase to provision and configure the
target OpenShift cluster before progressing to further steps. At this
point the role only creates the expected needed projects/namespaces for
image pushing and deployments, but it will contain RBAC configuration as
well.
This commit starts to build up a more generic and consolidated way of
mananing cluster commands through the Ansible module instead of direct
calls to OC and adds a global `cifmw_kubeconfig` variable that is
expected to be present always after the infra step and that any other
step further in the chain can use to access the cluster.