Refactor documentation and add new content
This patch changes a bit the doc structure in order to make it more
convenient for both accessing data, and maintaining content.

We also expose an FAQ, allowing people to consume the framework without
being blocked by known problem we can't easily solve.

We also take the opportunity to clean the README as well as USAGE, so
that everything is only in one location.

This patch also exposes the README listing the various ansible plugins
we have, in order to make them easier to understand.

Finally, we also take the opportunity to expose the built doc in the
zuul job, so that we actually know how it will displayed.