Create /home/zuul directory in job workspace
In EDPM content provider job, the content provider has zuul user
and edpm job has zuul-worker user. When content provider returns
meta operator git_src data to the EDPM job. It contains /home/zuul
user in the path. It breaks the openstack services deployment by saying
```
/home/zuul/src/github.com/openstack-k8s-operators/openstack-operator
directory not found on EDPM node.
```

We tried to set '~' for pointing to user's home[1] directory but ansible
still resolves to /home/zuul user.

In order to fix that, we are adding a pre-playbook to create a
/home/zuul directory and create a symlink to /home/zuul-worker. It will
fix the above issue.

[1]. https://github.com/openstack-k8s-operators/dataplane-operator/pull/206
Signed-off-by: Chandan Kumar <raukadah@gmail.com>