Update the PR template
This should hopefully raise awareness on the needs for a change to get
merged.

This also enforces checklist ticking, in the hope contributors and
reviewers will be more aware of the needs, especially related to the
documentation update.
The workflow has a strong requirement on the checklist created in the PR
template, and should fail if we don't included it, or if some ticks are
missing.

As a pull request owner and reviewers, I checked that:
- [ ] Appropriate testing is done and actually running
- [ ] ~~Appropriate documentation exists and/or is up-to-date:
  - [ ] ~~README in the role
  - [ ] ~~Content of the docs/source is reflecting the changes