Added new openshift_login role
This roles enables the framework to perform OpenShift logins in a
consistent way by centralizing all login related actions into it.
It based on explicit calls to the `oc` cli tool, that manages
authentication underneath based on a given set of role variables.
Depending on the variables that are passed to the role the `oc
login` command arguments are build allowing multiple scenarios, like:
- Login with user, password and API endpoint
- Login with a kubeconfig file only
- Login with a kubeconfig file but overriden user/password
Despite the given login variables used, the role sets at the end of it's
execution the API URL, user, context, kubeconfig and token that applies
to the whole run, taht way, any consumer in the playbook can use those
credentials to login.