Improve documentation related to the requirements
We must list the needed packages as well as the potential privileges we
need.
Mostly, we must point the `libvirt` group belonging.

A potential future improvement is to get a dedicated playbook that will
take care of the initial env bootstrap, to be launched only once. It
will check for the requirements, and output a nice message at the end,
asking the user to login/logout in order to refresh groups, for
instance, or any service message that may help ensuring a successful
deploy.