Remove crc bits from cifmw-base-crc-openstack
cifmw-base-crc-openstack uses base-simple-crc as a parent
and contains ci-framework and crc related jobs vars.

It does not give a clear seperation between crc and ci-framework.
This patch breaks the cifmw-base-crc-openstack into two jobs
1. cifmw-base-crc containing the crc bits
2. cifmw-base-crc-openstack containing the ci-framework bits with
   pre-run and post-run playbook calls and vars

cifmw-base-crc-openstack base job will provide the complete environment
to use the ci-framework starting from pre-setup to log collection.

The run playbook needs to be called in the main job as it does not
gets overriden when main job is used as a parent.

Note: cifmw_openshift_kubeconfig is already defined in scenario files,
so dropping it from zuul job definitions.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>