Add playbook to validate podified deployment
On install_yamls repo, we run podified control plane
and edpm based zuul jobs.

The existing ci-framework provides a way to deploy podified
and edpm deployment. It also provides utility to validate
the edpm deployment but not podified deployment.

This patch adds the playbook to validate the deployment and
improves the coverage against install_yamls.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>