Convert -nobuild job to -tagged-nobuild
This job goal is to ensure tags are covering the needs accordingly.

Running the deploy-edpm.yml playbook three times with a combination of
`--tags` and `--skip-tags` should ensure we're delivering the needed
feature.

Note that, as we get more and more tags, we may need to revisit this
base job.

Since it's a 100% overlap with the "-nobuild" job, we just convert that
one, saving resources.