Correct issues
While adding a new `end-to-end-tagged` job, we found some issues with
tasks that were relying on undefined parameters - mostly because the
task defining those parameters weren't ran due to tag filtering.

We also detected some missing "always" and "bootstrap" tags, so this
patch also adds them for the sake of consistency.