Adds cifmw_extras to kuttl-run playbook
This patch adds cifmw_extras to kuttl run playbook to allow passing
different scenarios files.
This patch also change the order of extra-vars in ansible-playbook
calls. By moving zuul_inventory.yml to the end, it will allow us to
overrides vars in zuul job definitions, which can be useful to:
 * avoid too many scenario files when change in the scenario is minimal
 * allow us to testproject a zuul job without changing the original scenario
   file.