Ensure we're having cifmw_path parameter
Since this playbook may be running outside of the main deploy-edpm.yml
play, we have to ensure cifmw_path is set.

This is especially true within zuul CI, since this 99-logs.yml is
launched from the post-run stage.

We could consume the generated custom-params.yml file, but since it may
not exist in case of (really) early failure of the "run" stage, we'd
rather statically add the missing parts.