Improve trigger for CRC log gathering
It may happen CRC is consumed without actually being bootstraped with
the Framework, meaning we may miss the `cifmw_use_crc` parameter.

This change tries to detect the CRC VM, re-using the rhol_crc role. In
order to do so, the specific code is now extracted in its own tasks
file, so that we may call it from anywhere without having the whole role
running.

We also introduce a "retry", while also ignore errors in case of final
failure.

The "retries" should give some more chances to the command to succeed,
since it may hit "Structure needs cleaning".