Increase CRC memory and CPU
Nodes have 32G, and we're hitting some resource shortage in the CRC VM.
Nodes have 16 core, so we can consume half of that for CRC.

We therefore also bump the e2e node flavor.

Note: It also decreases the bmass node count to 1 as the baremetal
was failing again with the existing memory bump.

Depends-On: https://github.com/openstack-k8s-operators/install_yamls/pull/332