Set DEPLOY_DIR var in BM jobs
https://github.com/openstack-k8s-operators/install_yamls/pull/339/files#diff-274a7c302ea3acca968d0bda33efaca145d8e14cb3b905c6f7a56b87fcd5ef38R18
adds DEPLOY_DIR var in BMAAS EDPM compute deployment.

This var needs to be set to collect proper logs and artifacts once
the deployment completes.

Note: DEPLOY_DIR var is used at multiple places during operator
deployment. It is not defined as a part of Makefile vars. So we
need to define it as a part of cifmw_install_yamls_whitelisted_vars
to make it working with ci-framework roles.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>