Pass environment variables to repo-setup cli
In Downstream, we use different dlrn url. In order to allow
repo-setup to access those urls, proper CA_CERTS environment
vars needs to be passed to repo-setup cli otherwise repo-setup
for downstream will fail.

This patch adds cifmw_repo_setup_env var to set the same.
It also modifies the artifacts task file to use command instead
of shell so that environment variables are applied to the repo-setup
cli.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>