Avoid "always" tag on install_yamls content
install_yamls role should be launched only in two cases:
- we add --tags boostrap
- we don't add any tag

Apparently, with the "always" being there, the tasks were launched even
though we wanted only the "packages" tag.

We still need to expose some of the generated facts, even if they are
empty, so that we can get a clean run with "--tags bootstrap".