Role to set meta operators env variables
This role provides a way to overrides the registry address of
all the meta operator env variables.

By default the role should update the address of openstack
services containers to different registry with proper tag.

It will be useful for downstream.

It also adds new molecule base based on crc.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>