Introduce Ansible roles and playbook for Ceph
These five roles and the ceph.yaml playbook can be used to
deploy Ceph on EDPM nodes for testing HCI.

The cifmw_cephadm, cifmw_ceph_client, and cifmw_create_admin
are from TripleO but only features needed for CI are included.
The cifmw_block_device role is from tripleo-ci and creates a
virtual block device, separate from the root volume, which can
be used as an OSD. The cifmw_ceph_spec role is introduced here
to render an opinionated Ceph spec for a CI job based on what
is found on EDPM nodes. The cifmw_create_admin role can be
used to create an SSH user for cephadm. Most of the work is
from the cifmw_cephadm role, which is an Ansible wrapper to
call the Ceph cephadm tool.

Signed-off-by: John Fulton <fulton@redhat.com>