Introduce new CI layout for end-to-end jobs
In order to improve stability while also allowing more flexibility with
"what we want to test", this patch introduces a new multinode layout for
CRC based end-to-end jobs.

cifmw-podified-edpm-base-multinode-crc provides the basis for the
3-nodes layout, providing one "ansible controller", a CRC instance and
one Compute node.
At the end of that base job, everything is ready to start running the CI
Framework - dependencies are installed, and you should get all of the
needed repositories, also for proper Depends-On support.

cifmw-podified-edpm-deployment-multinode-crc is an actual end-to-end
job, running the standard EDPM deployment with some tests.

We currently migrate the following jobs to the new layout:
- ci-framework-crc-podified-galera-deployment
  moved to cifmw-podified-galera-deployment-multinode-crc
- ci-framework-crc-podified-edpm-deployment
  moved to cifmw-podified-edpm-deployment-multinode-crc
- cifmw-end-to-end-nobuild-tagged
  moved to cifmw-podified-edpm-e2e-nobuild-tagged-multinode-crc

The following job is *removed* from the CI run:
cifmw-end-to-end
Reason: mostly a duplicate of the edpm-deployment-multinode-crc; we're
also waiting for a new job to fully replace it, introducing Ceph HCI.