Filter hosts in CI playbook
We cannot use a parameter as `hosts` value in zuul playbooks. The
parameters as passed via the "vars" and related job parameters are
injected in the inventory, and not exposed during the ansible-playbook
startup, but only once it's loaded and ready.

At that stage, the playbook already defaulted to the `all` value,
preventing clean and nice filtering.

We'd need to pass the parameter as actual --extra-vars, but this isn't
possible in zuul.

We therefore have to leverage the `end_host` meta to skip the host if we
set the cifmw_zuul_target_host parameter.