Improved the way we load install_yamls defaults
Till now we simply parse install_yamls Makefiles to search for variable
definitions but those variables assignations now contains conditionals
and shell calls that makes the current dumped values unreliable. The
idea with this commit is to make make itself handle the parse of the
variables. To acomplish that we create a copy of the Makefile in where
we add an autogenerated target that simply dumps all the declared vars
to a file that will be picked up and parsed to extract the computes
values.