Change base container image ubi9->centos
A FIPS enabled EDPM node is currently failing because libvirt doesn't
start[1], logging this message:

  error : virRandomBytes:113 : internal error: failed to generate byte
          stream: An error has been detected in the library and cannot continue
          operations.

In any FIPS enabled environment this issue can be replicated in a
container:

  # podman run --rm quay.io/podified-antelope-centos9/openstack-nova-libvirt:current-podified gnutls-cli-debug
  Error in GnuTLS initialization: Error while performing self checks.
  global state initialization error

The root cause of this is that the centos gnutls package installed in
the container doesn't work with a ubi9 base image when FIPS is enforced.
The centos gnutls works with a centos base image, and the ubi gnutls
package works with a ubi9 base image[2].

This change switches the base image to centos:stream9 specifically to
fix this FIPS issue, but there may be other problems with mixing
packages across distros.

[1] http://pastebin.test.redhat.com/1108025
[2] http://pastebin.test.redhat.com/1108270