Add Dev-preview workaround to finish EDPM deployment
It is a know issue with the dev-preview nova-operator that
it runs the libvirt and nova-compute deployment twice but
reports ready status after the first run. So with this sleep
the job waits for the deployment to really stabilize before
start running the tempest tests.

Adding 5 sleep before running tests as a workaround fixes
the issue.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>