Remove galera job
This job isn't really useful - it's better to add the parameter into the
edpm-deployment job instead.

We also add a small doc about nodeset override in the Zuul job
definition, since it may not be that obvious.

And we correct a small issue that may happen on log collection, when the
host doesn't have libvirt installed/running.