Deploy RGW as part of Ceph deployment
It's possible to configure swift with an external RGW backend.
This change allows the Ceph deploy to support a configured external
RGW instance that can serve swift related requests. The main ceph
playbook has been modified to include the rgw task, and a spec
template is rendered to deploy RGW via the ceph orchestrator.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>