Pin Ceph container to latest-reef
Reef has been released recently [1] and the upstream development is
going to see focus on the next "S" cycle. In order to avoid unexpected
behaviors, the idea is to pin the Ceph container version to the latest
Reef tag: this allow us to track regressions with the latest generated
content and stay on track with the target release for Antelope OpenStack
containers.

[1] https://ceph.io/en/news/blog/2023/v18-2-0-reef-released/

Signed-off-by: Francesco Pantano <fpantano@redhat.com>