Switch ssh key to our own custom one, create cloud-admin
It seems that "cloud-admin" is the user to use in a really close
future[1],so let's ensure it's present on the compute. Also, it
seems "root" is being used for now, and it's probably good to use
our own key, since we ensure it's present in all the needed users.

[1] https://github.com/openstack-k8s-operators/dataplane-operator/pull/398/files