Fix Ceph Spec play to correctly map hostname to storage IP
This patch updates the "Build a dict mapping hostname to its IP
which is in storage network range" task to correctly retrieve
the storage IP from the hostvars.

Fixes bug introduced by 2de93b92fc26f844b75681ddc7ec71f8279fb786

In breaking commit, `hostvars[item][all_addresses]` was changed
to `hostvars['ansible_hostname']` which does not contain the same
value. Runs of the play after then failed with "'bool' object is
not iterable".

Also, add `when` to prevent an undefined key in the dictionary.

Signed-off-by: John Fulton <fulton@redhat.com>